# Application/openapi.json
{
  "components": {
    "schemas": {
...
      "ComboIn": {
        "properties": {
          "combo_price_vnd": {
            "title": "Combo Price Vnd",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ComboItemIn"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "target_group": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Group"
          },
          "validity_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validity End"
          },
          "validity_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validity Start"
          }
        },
        "required": [
          "name",
          "combo_price_vnd"
        ],
        "title": "ComboIn",
        "type": "object"
      },
      "ComboItemIn": {
        "properties": {
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "default": 1,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "product_id"
        ],
        "title": "ComboItemIn",
        "type": "object"
      },
      "ComboItemOut": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_id": {
            "title": "Product Id",
            "type": "integer"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "product_id",
          "quantity",
          "name"
        ],
        "title": "ComboItemOut",
        "type": "object"
      },
      "ComboOut": {
        "properties": {
          "combo_id": {
            "title": "Combo Id",
            "type": "integer"
          },
          "combo_price_vnd": {
            "title": "Combo Price Vnd",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComboItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ComboStatus"
          },
          "target_group": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Group"
          },
          "validity_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validity End"
          },
          "validity_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validity Start"
          }
        },
        "required": [
          "combo_id",
          "name",
          "description",
          "combo_price_vnd",
          "target_group",
          "validity_start",
          "validity_end",
          "status",
          "items"
        ],
        "title": "ComboOut",
        "type": "object"
      },
      "ComboPatch": {
        "properties": {
          "combo_price_vnd": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combo Price Vnd"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ComboItemIn"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "target_group": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Group"
          },
          "validity_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validity End"
          },
          "validity_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validity Start"
          }
        },
        "title": "ComboPatch",
        "type": "object"
      },
      "ComboStatus": {
        "enum": [
          "Scheduled",
          "Active",
          "Expired"
        ],
        "title": "ComboStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PizzaHUST API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/combos": {
      "get": {
        "operationId": "list_combos_api_admin_combos_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComboOut"
                  },
                  "title": "Response List Combos Api Admin Combos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Combos",
        "tags": [
          "admin-combos"
        ]
      },
      "post": {
        "operationId": "create_combo_api_admin_combos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComboIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComboOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Combo",
        "tags": [
          "admin-combos"
        ]
      }
    },
    "/api/admin/combos/{combo_id}": {
      "delete": {
        "operationId": "delete_combo_api_admin_combos__combo_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "combo_id",
            "required": true,
            "schema": {
              "title": "Combo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Combo",
        "tags": [
          "admin-combos"
        ]
      },
      "get": {
        "operationId": "get_combo_api_admin_combos__combo_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "combo_id",
            "required": true,
            "schema": {
              "title": "Combo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComboOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Combo",
        "tags": [
          "admin-combos"
        ]
      },
      "patch": {
        "operationId": "patch_combo_api_admin_combos__combo_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "combo_id",
            "required": true,
            "schema": {
              "title": "Combo Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComboPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComboOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Combo",
        "tags": [
          "admin-combos"
        ]
      }
    },
...
  }
}